+++ bramble-java/src/main/java/org/briarproject/bramble/plugin/tor/JavaTorPlugin.java
@@ -33,7 +33,7 @@ abstract class JavaTorPlugin extends TorPlugin {
 			TorRendezvousCrypto torRendezvousCrypto,
 			PluginCallback callback,
 			String architecture,
-			int maxLatency,
+			long maxLatency,
 			int maxIdleTime,
 			File torDirectory) {
 		super(ioExecutor, wakefulIoExecutor, networkManager, locationUtils,
